function that returns 1 if the input integer is a prime number, otherwise return 0.

+++ 0x08-recursion/5-sqrt_recursion.c
@@ -9,7 +9,7 @@
  */
 
 
-int sqrt_num(int a, int b)
+int sqrt_num(long int a, long int b)
 {
 	if (a == 1)
 		return (1);
@@ -32,5 +32,4 @@ int _sqrt_recursion(int n)
 {
 
 	return (sqrt_num(n, n / 2));
-
 }
